gliner/requirements.txt
gliner>=0.2.0
pyyaml>=6.0
